+++ src/Router.js
@@ -5,7 +5,6 @@ import Main from './pages/Main/Main';
 import Product from './pages/Product/Product';
 import Detail from './pages/Product/Detail/Detail';
 import Footer from './components/Footer/Footer';
-import Cart from './pages/Cart/Cart';
 
 const Router = () => {
   return (
@@ -15,8 +14,6 @@ const Router = () => {
         <Route path="/" element={<Main />} />
         <Route path="/product" element={<Product />} />
         <Route path="/detail" element={<Detail />} />
-        {/* <Route path="/" element={<Nav />} />
-        <Route path="/" element={<Cart />} /> */}
       </Routes>
       <Footer />
     </BrowserRouter>
