+++ src/components/Drawer/Drawer.scss
@@ -7,15 +7,20 @@
   border-top: 1px solid #f5f5f5;
   font-size: 15px;
   background-color: white;
-}
-
-.subCategory {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(10%, auto));
-  margin-bottom: 15px;
-}
-
-.subCategory:first-child {
-  font-weight: bold;
-  color: #30761E;
+  .subCate {
+    display: inline-flex;
+    flex-direction: column;
+    margin: 30px 50px;
+    .menuCate {
+      margin-bottom: 10px;
+    }
+    .menuCate:first-child {
+      font-weight: bold;
+      margin-bottom: 10px;
+      color: #30761e;
+    }
+  }
+  .subCate:first-child {
+    margin-left: 180px;
+  }
 }
